add upstream fixes to entrypoint

+++ entrypoint.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+et -e
+
 # Install netlify globally before NVM to prevent EACCESS issues
 npm i -g netlify-cli
 
@@ -7,10 +9,17 @@ npm i -g netlify-cli
 NETLIFY_CLI=$(which netlify)
 
 # Install node from NVM to honor .nvmrc files
-curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.35.2/install.sh | bash
-[ -s "$HOME/.nvm/nvm.sh" ] && \. "$HOME/.nvm/nvm.sh"
+if [[ -n $9 ]] || [[ -e ".nvmrc" ]]; then
+	curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.35.2/install.sh | bash
+	[ -s "$HOME/.nvm/nvm.sh" ] && \. "$HOME/.nvm/nvm.sh"
 
-nvm install "$9"
+	nvm install "$9"
+	if [[ -n $9 ]]; then
+		nvm install "$9"
+	else
+		nvm install
+	fi
+fi
 
 NETLIFY_AUTH_TOKEN=$1
 NETLIFY_SITE_ID=$2
